fix: Handle parsing of temps between -10 and 10°C
Data returned from the board is in the form of
"tt1613,ht37" (top board, 16.13 deg C, RH 37%).
Parsing was done by counting characters, expecting the temperature in
the [2:6] range. For single digits (ie TT314,ht37), this included the
comma.
Parsing is now done by splitting on the comma and discarding the first
two characters, leaving only the integer part of the string.